+++ src/revlanguage/help/IHelp.h
@@ -14,7 +14,7 @@
 
 class IHelp {
 public:   
-    virtual std::string getHelp(std::string query, int columnWidth) = 0;    
+    virtual std::string getHelp(std::string query, size_t columnWidth) = 0;
     virtual bool isHelpAvailableForQuery(std::string query) = 0;
     
     void setRbHelpEntry(RbHelpEntry rbHelp) {
